+++ platformio/project/options.py
@@ -697,19 +697,6 @@ ProjectOptions = OrderedDict(
                     "and returns results to the standard output"
                 ),
             ),
-            ConfigEnvOption(
-                group="test",
-                name="test_verbosity_level",
-                description=(
-                    "Verbosity level: "
-                    "0=normal verbosity (default), "
-                    "1=raw testing output, "
-                    "2=base verbosity for buidling/uploading, "
-                    "3=extra verbosity for building/uploading"
-                ),
-                type=click.IntRange(min=0, max=3),
-                default=0,
-            ),
             # Debug
             ConfigEnvOption(
                 group="debug",